Increased Base & Added Comment Filtering
If a comment was removed on the subreddit it was posted to by moderators
then those comments would still be counted since the posts were gathered
from the user's profile. There is now a button to have each post
cross-checked with its status on the subreddit to know if it needs to
filter those out.